requirements.txt
sanic[ext]
sanic-session
toml
tortoise-orm
pillow
aiomysql